Skeleton of the runtime mostly ready.

+++ runtime/src/mm/mmap.c
@@ -4,7 +4,7 @@
 
 void *mmap(void *start, size_t len, int prot, int flags, int fd, off_t off)
 {
-  long rs = __syscall6(SYS_mmap, start, len, prot, flags, fd, off);
+  long rs = __syscall6(SYS_mmap, (long)start, len, prot, flags, fd, off);
   if (rs < 0) {
     errn = (unsigned) -rs;
     return MAP_FAILED;
